feat: space main 조회 추가

+++ src/main/java/one/dfy/bily/api/inquiry/dto/InquiryResponse.java
@@ -21,7 +21,7 @@ public record InquiryResponse(
         LocalDateTime createdAt,
         InquiryStatus status,
         String author,
-        Integer spaceId,
+        Long spaceId,
         String hostCompany,
         String spaceIdName
 ) {}
